UI
Bind data from sensor and API to the main page UI.

+++ app/src/main/java/com/example/homeassistant/domain/api/AirPollution.kt
@@ -6,12 +6,12 @@ data class AirPollution(
     val location: Location,
     val dateTime: Long,
     val airQualityIndex: Int,
-    val carbonMonoxide: Float,
-    val nitrogenMonoxide: Float,
-    val nitrogenDioxide: Float,
-    val ozone: Float,
-    val sulphurDioxide: Float,
-    val fineParticles: Float,
-    val coarseParticles: Float,
-    val ammonia: Float
+    val carbonMonoxide: Double,
+    val nitrogenMonoxide: Double,
+    val nitrogenDioxide: Double,
+    val ozone: Double,
+    val sulphurDioxide: Double,
+    val fineParticles: Double,
+    val coarseParticles: Double,
+    val ammonia: Double
 )
