+++ app/src/main/java/com/example/homeassistant/domain/api/Forecast.kt
@@ -2,16 +2,16 @@ package com.example.homeassistant.domain.api
 
 data class Forecast(
     val dateTime: Long,
-    val temperature: Float,
-    val feelsLike: Float,
-    val minTemperature: Float,
-    val maxTemperature: Float,
+    val temperature: Double,
+    val feelsLike: Double,
+    val minTemperature: Double,
+    val maxTemperature: Double,
     val pressure: Int,
     val humidity: Int,
     val mainWeather: String,
     val description: String,
     val cloudiness: Int,
-    val windSpeed: Float,
+    val windSpeed: Double,
     val visibility: Int,
-    val precipitation: Float
+    val precipitation: Double
 )
